Integrate authentication with frontend
