--- NEMRA-Apartment/SE-group-project-group_12_nemra-apartment_management_system/Apartment-Management-System/backend/src/main/java/com/AMS/Apartment_Management_System/ApartmentManagementSystemApplication.java
@@ -1,13 +1,0 @@
-package com.AMS.Apartment_Management_System;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ApartmentManagementSystemApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ApartmentManagementSystemApplication.class, args);
-	}
-
-}
