--- NEMRA-Apartment/SE-group-project-group_12_nemra-apartment_management_system/Apartment-Management-System/backend/src/main/java/com/AMS/Apartment_Management_System/controllers/LoginController.java
@@ -1,45 +1,0 @@
-package com.AMS.Apartment_Management_System.controllers;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.AMS.Apartment_Management_System.Services.UserServiceImpl;
-
-@RestController
-public class LoginController {
-
-	private final UserServiceImpl userService;
-
-    @Autowired
-    public LoginController(UserServiceImpl userService) {
-        this.userService = userService;
-    }
-	
-	@RequestMapping("/login")
-	public String showLoginPage() {
-		return "index.html"; 
-	}
-
-	@PostMapping("/login")
-	public String processLogin(@RequestParam("username") String username, 
-                               @RequestParam("password") String password,
-							   Model model) {
-		
-		System.out.println("username: " + username);
-		System.out.println("password: " + password);
-		
-		if (userService.validateAuthentication(username, password)) {
-			System.out.println("Valid username and password");
-			return "successful";
-		}
-		else {
-			System.out.println("Invalid username or password");
-			return "successful";
-		}
-	}
-}
