--- NEMRA-Apartment/SE-group-project-group_12_nemra-apartment_management_system/Apartment-Management-System/backend/src/main/resources/application.properties
@@ -1,14 +1,0 @@
-
-# Database connection properties
-spring.datasource.url=jdbc:postgresql://localhost:5432/nemra
-spring.datasource.username=postgres
-spring.datasource.password=SE2072
-spring.jpa.hibernate.ddl-auto=update
-#spring.jpa.hibernate.ddl-auto=<create | create-drop | update | validate | none>
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
-#spring.liquibase.clear-checksums=true
-server.port=8081
-#spring.thymeleaf.check-template-location=false
-
-
-#5432
