--- NEMRA-Apartment/SE-group-project-group_12_nemra-apartment_management_system/Apartment-Management-System/backend/src/test/java/com/AMS/Apartment_Management_System/ApartmentManagementSystemApplicationTests.java
@@ -1,13 +1,0 @@
-package com.AMS.Apartment_Management_System;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApartmentManagementSystemApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
