--- NEMRA-Apartment/SE-group-project-group_12_nemra-apartment_management_system/Apartment-Management-System/frontend/vite.config.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-
-// https://vite.dev/config/
-export default defineConfig({
-  plugins: [react()],
-})
